requirements.txt
bleach==1.5.0
cycler==0.10.0
decorator==4.1.2
enum34==1.1.6
h5py==2.7.1
html5lib==0.9999999
Keras==2.0.9
Markdown==2.6.9
matplotlib==2.1.0
networkx==2.0
numpy==1.13.3
olefile==0.44
Pillow==4.3.0
protobuf==3.4.0
pygame==1.9.3
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyWavelets==0.5.2
PyYAML==3.13
scikit-image==0.13.1
scipy==1.0.0
six==1.11.0
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0rc2
Werkzeug==0.12.2
pytorch
torchvision